chore: move `zipAndDeleteModuleWithTest` inside zipUtils.ts
